separate components and functions to file
