+++ summing-grid/src/func.js
@@ -1,0 +1,1 @@
+export const summing = (arr) => arr.reduce((sum, ele) => sum + Number.parseFloat(ele || 0), 0)
